requirements.txt
# utility
tqdm==4.66.1
scikit-image==0.21.0
imageio==2.31.1
Pillow==10.2.0
numpy==1.24.1
natsort==8.4.0

# EXR handling 
skylibs==0.7.4
OpenEXR==1.3.9

# We use pytorch pip instead because conda is mess up
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.1.1+cu121
torchvision==0.16.1+cu121
torchaudio==2.1.1+cu121

# Diffusers dependencies
accelerate==0.21.0
datasets==2.13.1
diffusers==0.21.0
transformers==4.36.0
xformers==0.0.23
huggingface_hub==0.19.4
